pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "glyphgen"
version = "0.1.0"
description = "Structured component-and-layout conditioned glyph generation."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "bb" }
]
dependencies = [
  "numpy>=1.26",
  "Pillow>=10.0",
  "PyYAML>=6.0",
  "tqdm>=4.66",
  "gradio>=4.0",
  "torch>=2.2",
  "torchvision>=0.17",
  "lpips>=0.1.4",
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

